refactor: share OpenAI Agents telemetry helpers
