+++ aws-opentelemetry-distro/src/amazon/opentelemetry/distro/instrumentation/openai_agents/_gen_ai_context_capture.py
@@ -1,0 +1,6 @@
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from ._wrappers import OpenAIAgentsWrappers
+
+__all__ = ["OpenAIAgentsWrappers"]
